requirements.txt
autopep8==1.5.7
backcall==0.2.0
click==8.0.1
colorama==0.4.4
decorator==5.0.9
et-xmlfile==1.1.0
Flask==2.0.1
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
numpy==1.20.3
openpyxl==3.0.7
pandas==1.2.4
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
pycodestyle==2.7.0
Pygments==2.9.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
pywin32==300
pyzmq==22.0.3
six==1.16.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==2.0.1
xlrd==2.0.1
